requirements.txt
flask==3.1.0
loguru==0.7.3
gitpython==3.1.44
git-filter-repo==2.47.0
